+++ packages/flutter_spanned_controller/CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `ToggleStateListener`, which is a generalization of `AttributeListener`
+
 ## [0.4.0-0] — 2020-12-22
 
 ### Removed